+++ src/boxrec-pages/profile/boxrec.page.profile.bout.row.ts
@@ -1,13 +1,4 @@
-import {
-    BoxingBoutOutcome,
-    BoxrecBasic,
-    BoxrecBout,
-    BoxrecBoutLocation,
-    BoxrecJudge,
-    BoxrecId,
-    Record,
-    WinLossDraw
-} from "../boxrec.constants";
+import {BoxrecBout, BoxrecBoutLocation} from "../boxrec.constants";
 import {trimRemoveLineBreaks} from "../../helpers";
 import {BoxrecCommonTablesClass} from "../boxrec-common-tables/boxrec-common-tables.class";
 
@@ -17,18 +8,8 @@ let $: CheerioAPI;
 export class BoxrecPageProfileBout extends BoxrecCommonTablesClass {
 
     private _date: string;
-    private _firstBoxerWeight: string;
-    private _opponent: string;
-    private _secondBoxerWeight: string;
-    private _opponentRecord: string;
-    private _opponentLast6: string;
-    private _outcome: string;
-    private _outcomeByWayOf: string;
-    private _rating: string;
     private _links: string;
     private _location: string;
-    private _numberOfRounds: string;
-    private _metadata: string;
 
     constructor(boxrecBodyBout: string, additionalData: string | null = null) {
         super();
@@ -43,10 +24,10 @@ export class BoxrecPageProfileBout extends BoxrecCommonTablesClass {
         return {
             date: this.date,
             firstBoxerWeight: this.firstBoxerWeight,
+            secondBoxer: this.secondBoxer,
             secondBoxerWeight: this.secondBoxerWeight,
-            opponent: this.opponent,
-            opponentRecord: this.opponentRecord,
-            opponentLast6: this.opponentLast6,
+            secondBoxerRecord: this.secondBoxerRecord,
+            secondBoxerLast6: this.secondBoxerLast6,
             titles: this.titles,
             referee: this.referee,
             judges: this.judges,
@@ -61,134 +42,6 @@ export class BoxrecPageProfileBout extends BoxrecCommonTablesClass {
     get date() {
         const date = this._date;
         return date.trim();
-    }
-
-    get firstBoxerWeight(): number | null {
-        return super.parseWeight(this._firstBoxerWeight);
-    }
-
-    get opponent(): BoxrecBasic {
-        const html = $(`<div>${this._opponent}</div>`);
-        const href: string = html.find("a").attr("href");
-        let opponent: BoxrecBasic = {
-            id: null,
-            name: null,
-        };
-        const regex: RegExp = /(\d+)$/;
-
-        if (href) {
-            const matches: RegExpMatchArray | null = href.match(regex);
-
-            if (matches && matches[1]) {
-                opponent = {
-                    id: parseInt(matches[1], 10),
-                    name: html.find("a").text(),
-                };
-            }
-        }
-
-        return opponent;
-    }
-
-    get referee(): BoxrecBasic {
-        const html = $(`<div>${this._metadata}</div>`);
-        let referee: BoxrecBasic = {
-            id: null,
-            name: null,
-        };
-
-        html.find("a").each((index: number, elem: CheerioElement) => {
-            const href: string = $(elem).get(0).attribs.href;
-            const matches: RegExpMatchArray | null = href.match(/referee\/(\d+)$/);
-
-            if (matches && matches[1]) {
-                const id: number = parseInt(matches[1], 10);
-                let name: string = $(elem).text();
-                name = trimRemoveLineBreaks(name);
-
-                referee.id = id;
-                referee.name = name;
-            }
-        });
-
-        return referee;
-    }
-
-    get judges(): BoxrecJudge[] {
-        const html = $(`<div>${this._metadata}</div>`);
-        let judges: BoxrecJudge[] = [];
-
-        html.find("a").each((index: number, elem: CheerioElement) => {
-            const href: string = $(elem).get(0).attribs.href;
-            const matches: RegExpMatchArray | null = href.match(/judge\/(\d+)$/);
-
-            if (matches && matches[1]) {
-                const id: number = parseInt(matches[1], 10);
-                let name: string = $(elem).text();
-                name = trimRemoveLineBreaks(name);
-
-                const scoreCardRegex = /<\/a>\s(\d{1,3})-(\d{1,3})/g;
-                let scorecard: number[] = [];
-
-                let m;
-                let i = 0;
-                do {
-                    m = scoreCardRegex.exec(this._metadata);
-                    if (m && m[1] && m[2]) {
-                        if (i === judges.length) {
-                            scorecard.push(parseInt(m[1], 10), parseInt(m[2], 10));
-                        }
-                    }
-                    i++;
-                } while (m);
-
-                judges.push({
-                    id, name, scorecard,
-                });
-            }
-        });
-
-        return judges;
-    }
-
-    get titles(): BoxrecId[] {
-        let titles: BoxrecId[] = [];
-        const html = $(`<div>${this._metadata}</div>`);
-        html.find("a.titleLink").each((index: number, elem: CheerioElement) => {
-            const href: string = $(elem).get(0).attribs.href;
-            const matches: RegExpMatchArray | null = href.match(/(\d+\/\w+)$/);
-
-            if (matches && matches[1]) {
-                const id: string = matches[1];
-                let name: string = $(elem).text();
-                name = trimRemoveLineBreaks(name);
-
-                titles.push({
-                    id, name,
-                });
-            }
-        });
-
-        return titles;
-    }
-
-    // maybe there's additional things that people would want to sift through
-    get metadata() {
-        return this._metadata;
-    }
-
-    get numberOfRounds(): (number | null)[] {
-        let numberOfRounds: (number | null)[] = [null, null];
-
-        const splitRounds: string[] = this._numberOfRounds.trim().split("/");
-        const formattedSplitRounds: number[] = splitRounds.map(item => parseInt(item, 10));
-        if (formattedSplitRounds.length === 2) {
-            numberOfRounds = formattedSplitRounds;
-        } else if (formattedSplitRounds.length === 1 && !isNaN(formattedSplitRounds[0])) {
-            numberOfRounds = [null, formattedSplitRounds[0]];
-        }
-
-        return numberOfRounds;
     }
 
     get location(): BoxrecBoutLocation {
@@ -253,26 +106,6 @@ export class BoxrecPageProfileBout extends BoxrecCommonTablesClass {
         return location;
     }
 
-    get secondBoxerWeight(): number | null {
-        return super.parseWeight(this._secondBoxerWeight);
-    }
-
-    get opponentRecord(): Record {
-        return super.parseRecord(this._opponentRecord);
-    }
-
-    get opponentLast6(): WinLossDraw[] {
-        return super.parseLast6Column(this._opponentLast6);
-    }
-
-    get result(): [WinLossDraw, BoxingBoutOutcome | string, BoxingBoutOutcome | string] {
-        return [this.outcome, this.outcomeByWayOf(), this.outcomeByWayOf(true)];
-    }
-
-    get rating(): number | null {
-        return super.parseRating(this._rating);
-    }
-
     // returns an object with keys that contain a class other than `clickableIcon`
     get links(): any { // object of strings
         const html = $(this._links);
@@ -301,44 +134,6 @@ export class BoxrecPageProfileBout extends BoxrecCommonTablesClass {
         return obj;
     }
 
-    get outcome(): WinLossDraw {
-        let outcome: string = this._outcome;
-        outcome = outcome.trim();
-        let formattedOutcome: WinLossDraw;
-
-        switch (outcome) {
-            case "W":
-                formattedOutcome = WinLossDraw.win;
-                break;
-            case "D":
-                formattedOutcome = WinLossDraw.draw;
-                break;
-            case "L":
-                formattedOutcome = WinLossDraw.loss;
-                break;
-            case "S":
-                formattedOutcome = WinLossDraw.scheduled;
-                break;
-            default:
-                formattedOutcome = WinLossDraw.unknown;
-        }
-
-        return formattedOutcome;
-    }
-
-    private outcomeByWayOf(parseText = false): BoxingBoutOutcome | string {
-        const html = $(`<div>${this._outcomeByWayOf}</div>`);
-        html.find("div").remove();
-        let outcomeByWayOf: string = html.text();
-        outcomeByWayOf = outcomeByWayOf.trim();
-
-        if (parseText) {
-            return BoxingBoutOutcome[outcomeByWayOf as any];
-        }
-
-        return outcomeByWayOf;
-    }
-
     private parseBout(): void {
         const getColumnData = (nthChild: number, returnHTML: boolean = true): string => {
             const el: Cheerio = $(`tr:nth-child(1) td:nth-child(${nthChild})`);
@@ -353,10 +148,10 @@ export class BoxrecPageProfileBout extends BoxrecCommonTablesClass {
         this._date = getColumnData(2, false);
         this._firstBoxerWeight = getColumnData(3, false);
         // empty 4th column
-        this._opponent = getColumnData(5);
+        this._secondBoxer = getColumnData(5);
         this._secondBoxerWeight = getColumnData(6, false);
-        this._opponentRecord = getColumnData(7);
-        this._opponentLast6 = getColumnData(8);
+        this._secondBoxerRecord = getColumnData(7);
+        this._secondBoxerLast6 = getColumnData(8);
         this._location = getColumnData(9);
         this._outcome = getColumnData(10, false);
         this._outcomeByWayOf = getColumnData(11);